VetScope/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
tensorflow-cpu==2.13.0
Pillow==10.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
numpy==1.24.3
python-dotenv==1.0.0 